requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
frozenlist==1.4.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
mysql==0.0.3
mysqlclient==2.2.4
openai==0.28.0
playsound==1.2.2
PyAudio==0.2.14
pydantic==2.7.0
pydantic_core==2.18.1
requests==2.31.0
sniffio==1.3.1
SpeechRecognition==3.10.3
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
